refactored unnecessary files
