+++ lib/board.js
@@ -1,7 +1,7 @@
 var $ = require("jquery");
 var rowFor = require("./row-for");
 var columnFor = require("./column-for");
-var Cell = require("./cell");
+
 var score = 0;
 var highScore = 0;
 
@@ -33,6 +33,12 @@ function Board() {
   this.column = 4;
   this.cells = [];
   this.buildGrid(arguments);
+}
+
+function Cell(row, column, value) {
+  this.row = row;
+  this.column = column;
+  this.value = value;
 }
 
 function Game () {
@@ -249,7 +255,7 @@ Cell.prototype = {
 };
 
 function lose() {
- alert('you lose');
+ alert('You lost. Play again!');
 }
 
 
@@ -273,4 +279,4 @@ function resetScore() {
 }
 
 
-module.exports = Board;
+module.exports = [Board, Cell];
