--- lib/game-manager.js
@@ -1,22 +1,0 @@
-var $ = require('jquery');
-var Board = require("../lib/board.js");
-
-
-
-function gameManager() {
-  // create new board with 16 cells
-
-//set up
-  // call Board.prototype.buildGrid to biuld grid and create object
-
-
-//startgame
-    // look for available cells with null value (checkAvailalbeCells)
-    // call changeValue/addTile to give 2 the value of 2
-    // render the page
-
-//gameover
-
-//restart game
-
-}
