+++ test/index.js
@@ -1,4 +1,3 @@
 require ("./board-test");
 require("./column-for-test");
 require("./row-for-test");
-require("./cell-test");
